+++ src/app/component-library/conditions-for-use-dialog/conditions-for-use-dialog/conditions-for-use-dialog.data.ts
@@ -1,0 +1,6 @@
+export interface ConditionsForUseDialogData {
+  title: string;
+  subtitle: string;
+  icon: string;
+  description?: string;
+}