+++ src/app/component-library/markdown-description/markdown-description.component.ts
@@ -11,7 +11,7 @@ import {
 } from '@angular/core';
 import {HtmlSanitizer} from 'src/app/core/services/html-sanitizer';
 import {MarkdownConverter} from 'src/app/core/services/markdown-converter';
-import {SimpleChangesTyped} from '../../../../core/utils/angular-utils';
+import {SimpleChangesTyped} from '../../core/utils/angular-utils';
 
 const COLLAPSED_DESCRIPTION_HEIGHT = 280;