+++ src/app/component-library/property-grid/property-grid.module.ts
@@ -5,6 +5,7 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 import {MatTooltipModule} from '@angular/material/tooltip';
+import {MarkdownDescriptionModule} from '../markdown-description/markdown-description.module';
 import {PipesAndDirectivesModule} from '../pipes-and-directives/pipes-and-directives.module';
 import {PolicyEditorModule} from '../policy-editor/policy-editor.module';
 import {UiElementsModule} from '../ui-elements/ui-elements.module';
@@ -29,6 +30,7 @@ import {PropertyGridComponent} from './property-grid/property-grid.component';
     PipesAndDirectivesModule,
     UiElementsModule,
     PolicyEditorModule,
+    MarkdownDescriptionModule,
   ],
   declarations: [PropertyGridComponent, PropertyGridGroupComponent],
   exports: [PropertyGridComponent, PropertyGridGroupComponent],